RL.GPI

Description

A general policy iterator for reinforcement learning.

Developed while doing the exercises in Ch. 4-6 of the book Reinforcement Learning: an Introduction, Richard S. Sutton and Andrew G. Barto, The MIT Press, Cambridge, Massachusetts; London, England

Documentation

data HypParams r

Hyper-parameters

The stated intents of the various fields are correct for the optPol function, which was the first solver written and performs dynamic programming (DP).

Some fields have overloaded meanings/uses, depending upon whether we're optPoling, optQing, or ... Individual functions below call out such differences in their header commentary as appropriate.

The type parameter r should be equal to the reward type of the MDP being solved.

Constructors

HypParams

Fields

disc :: r

discount rate

epsilon :: r

convergence tolerance

alpha :: r

error correction gain

beta :: r

epsilon/alpha decay factor

maxIter :: Int

max. value function evaluation iterations

tdStepType :: TDStep

temporal difference step type

nSteps :: Int

# of steps in n-step TD

mode :: Mode

optimization mode

hypParamsDef :: HypParams Double

A default value of type HypParams, to use as a starting point.

data TDStep

Type of temporal difference (TD) step.

Constructors

Sarsa

Qlearn

ExpSarsa

data Mode

Optimization mode.

Constructors

MC

TD

data Dbg s a r

Abstract type for capturing debugging information.

Constructors

Dbg

Fields

curSt :: s

act :: a

nxtSt :: s

rwd :: r

nxtStPrbs :: [(s, Double)]

data TDRetT s a r m n

Abstract type of return value from doTD function.

Constructors

TDRetT

Fields

valFuncs :: [s -> r]

polFuncs :: [s -> a]

polXCnts :: [Int]

valErrs :: [r]

debugs :: [[Dbg s a r]]

qMats :: [Vector m (Vector n (Double, Int))]

doTD

Arguments

:: (MDP s, Ord s, Eq (ActionT s), HasFin' s, HasFin' (ActionT s))

=> HypParams Double
Simulation hyper-parameters.

-> Int
Number of episodes to run.

-> TDRetT s (ActionT s) Double (Card s) (Card (ActionT s))

Find optimum policy and value functions, using temporal difference.

This is intended to be the function called by client code wishing to perform temporal difference based policy optimization.

type OptQT s g = (s, Vector (Card s) (Vector (Card (ActionT s)) (Double, Int)), g, [Dbg s (ActionT s) Double], Integer)

Object of optQn function.

Tuple element descriptions:

Current state.

Augmented matrix representation of Q(s,a). Cell type: (value, # of adjustments).

Random number generator.

List of debugging data structures.

Simulation time in units of state transitions.

type Unop a = a -> a

optQn

Arguments

:: (MDP s, HasFin' s, Ord s, Eq (ActionT s), HasFin' (ActionT s), RandomGen g)

=> HypParams Double
Simulation hyper-parameters.

-> Unop (OptQT s g)

Yields a single episodic action-value improvement iteration, using n-step TD.

HypParams field overrides:

epsilon: Used to form an "epsilon-greedy" policy.

This function accomodates the Monte Carlo (MC) method by using the following method of mean calculation, which has the same form as temporal difference (TD) error correction:

\[ y = f(x) \z = g(y)

\]

optPol

Arguments

:: (MDP s, Eq s, Ord s, HasTrie s, Ord (ActionT s), HasTrie (ActionT s))

=> HypParams Double
Simulation hyper-parameters.

-> (s -> (ActionT s, Double), [Int])
initial policy & value functions

-> (s -> (ActionT s, Double), [Int])

Yields a single policy improvment iteration.

HypParams field overrides:

maxIter: max. policy evaluation iterations (0 = Value Iteration)

Returns a combined policy & value function.

runEpisode

Arguments

:: Eq s

=> Int
Max. state transitions

-> (s -> a)
Policy

-> (s -> a -> s)
Next state calculator

-> [s]
Terminal states

-> s
Initial state

-> [s]

Run one episode of a MDP, returning the list of states visited.

Orphan instances

KnownNat n => HasTrie (Finite n)

Associated Types

data (Finite n) :->: b :: *

Methods

trie :: (Finite n -> b) -> Finite n :->: b

untrie :: (Finite n :->: b) -> Finite n -> b

enumerate :: (Finite n :->: b) -> [(Finite n, b)]