RL.MDP

Description

Markov Decision Process (MDP)

Documentation

class MDP s where

Markov Decision Process (MDP) (finite discrete)

Laws:

jointPMF must return a true probability distribution. That is:

 1 = sum . map snd $ jointPMF s a

\(, \forall s \in \text{states}, \forall a \in \text{actions s}\)

Usage notes:

If initStates is empty then use head states as the initial state; else randomly select initial state from initStates.

The notation used in the commentary comes from the Sutton & Barto text.

Minimal complete definition

states, actions, jointPMF

Associated Types

type ActionT s :: *

Methods

states :: [s]

State enumeration function - \(S\).

actions :: s -> [ActionT s]

Action enumeration function - \(A(s)\).

jointPMF :: s -> ActionT s -> [((s, Double), Double)]

Joint probability distribution - \(Pr[(s', r) | s, a]\).

initStates :: [s]

Initial states.

termStates :: [(s, Double)]

Terminal states and their values.

nextStates :: (MDP s, Eq s, Ord s) => s -> ActionT s -> [(s, Double)]

Next states and their probabilities - S'(s, a).

rewards :: (MDP s, Eq s) => s -> ActionT s -> s -> [(Double, Double)]

Rewards and their probabilities - R(s, a, s').

combProb :: (Eq a, Ord a) => [(a, Double)] -> [(a, Double)]

Eliminate duplicates from a probability distribution, by combining like terms and summing their probabilities.