RL.Util

Description

Common utilities used throughout the RL package.

Documentation

nat :: forall n. KnownNat n => Integer

Convenient abbreviation of: `natVal (Proxy @n)`.

int :: forall n. KnownNat n => Int

Convenient abbreviation of: `fromIntegral (nat @n)`.

appFm

Arguments

:: (HasFin' a, HasFin' b)

=> Vector (Card a) (Vector (Card b) c)
matrix representation of f(r,c)

-> a
row

-> b
column

-> c

Apply the matrix representation of a two argument function.

The first argument is assumed to index the rows of the matrix.

appFv

Arguments

:: HasFin' x

=> Vector (Card x) a
vector representation of f(x)

-> x
function argument

-> a

Apply the vector representation of a function.

qToV

Arguments

:: (Ord a, KnownNat m, KnownNat n, KnownNat k, n ~ (k + 1))

=> Vector m (Vector n a)
matrix representation of Q(s,a)

-> Vector m a
vector representation of V(s)

Convert an action-value matrix to a value vector.

(i.e. - \(Q(s,a) -> V(s)\))

qToP

Arguments

:: (Ord a, HasFin' act, KnownNat m, KnownNat n, KnownNat k, n ~ (k + 1), Card act ~ n)

=> Vector m (Vector n a)
matrix representation of Q(s,a)

-> Vector m act
vector representation of P(s)

Convert an action-value matrix to a policy vector.

(i.e. - \(Q(s,a) -> A(s)\))

newtype Pfloat

To control the formatting of printed floats in output matrices.

Constructors

Pfloat

Fields

unPfloat :: Float

newtype Pdouble

To control the formatting of printed doubles in output matrices.

Constructors

Pdouble

Fields

unPdouble :: Double

poisson :: Finite 5 -> Finite 12 -> Float

fact :: Int -> Int

poissonVals :: Vector 5 (Vector 12 Float)

poisson' :: Finite 5 -> Finite 21 -> Float

gammaPdf :: Double -> Double -> Double -> Double

Gamma pdf

Assuming `scale = 1`, shape should be: 1 + mean.

gamma :: Finite 5 -> Finite 12 -> Double

Gamma pmf

Scale assumed to be `1`, so as to match the calling signature of poisson.

gammaVals :: Vector 5 (Vector 12 Double)

gamma' :: Finite 5 -> Finite 21 -> Double

withinOnM :: Monad m => Double -> (a -> m Double) -> [a] -> m (Maybe a)

Monadically search list for first element less than given threshold under the given function, and return the last element if the threshold was never met. Return Nothing if the input list was empty.

withinIxM :: Monad m => Double -> [m Double] -> m (Maybe Int)

Monadically find index of first list element less than or equal to given threshold, or the index of the last element if the threshold was never met.

A return value of Nothing indicates an empty list was given.

maxAndNonZero :: (Foldable t, Num a, Ord a) => t a -> Writer [Int] a

Return the maximum value of a set, as well as a count of the number of non-zero elements in the set.

(See documentation for chooseAndCount function.)

chooseAndCount

Arguments

:: (Foldable t, Num a)

=> (a -> a -> a)
choice function

-> (a -> Bool)
counting predicate

-> t a
foldable set of elements to count/compare

-> Writer [Int] a

Choose a value from the set using the given comparison function, and provide a count of the number of elements in the set meeting the given criteria.

for :: [a] -> (a -> b) -> [b]

vsFor :: Vector n a -> (a -> b) -> Vector n b

mean :: (Foldable f, Fractional a) => f a -> a

Mean value of a collection

arrMeanSqr :: (Functor f, Foldable f, Functor g, Foldable g, Fractional a) => f (g a) -> a

Find the mean square of a list of lists.

sqr :: Num a => a -> a

Take the square of a numerical type.

toString :: (Show a, Typeable a) => a -> String

Convert any showable type to a string, avoiding the introduction of extra quotation marks when that type is a string to begin with.

boolToDouble :: Bool -> Double

Convert a Bool to a Double.

takeEvery :: Int -> [a] -> [a]

Take every nth element from a list.